+++ examples/selenium/testng/src/main/java/example/TestNG.java
@@ -1,0 +1,29 @@
+package example;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+
+public class TestNG {
+
+    private WebDriver driver;
+
+    @BeforeClass
+    public void setUp() {
+        driver = new ChromeDriver();
+    }
+
+    @Test
+    public void testOpenUrl() {
+        driver.get("https://www.blazedemo.com");
+    }
+
+    @AfterClass
+    public void tearDown() {
+        if (driver != null) {
+            driver.quit();
+        }
+    }
+}